文件映射1(版本合并)

+++ fms-domain/src/main/java/com/fms/domain/columnSet/ColumnInfo.java
@@ -10,7 +10,7 @@ public class ColumnInfo {
     private int isDic;
     private int isMasterKey;
     private String dataType;
-    private int tableId;
+    private long tableId;
     private int schemaId;
     private String dicTableName;
     private String columnChinese;
